django_backend/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
crispy-bootstrap4==2024.1
Django==5.0.9
django-ckeditor-5==0.2.13
django-crispy-forms==2.3
django-js-asset==2.2.0
django-mssql-backend==2.8.1
django-pyodbc==1.1.3
dnspython==2.7.0
idna==3.10
mssql-django==1.5
mysqlclient==2.2.4
numpy==2.1.1
packaging==24.1
pandas==2.2.2
pillow==10.4.0
plotly==5.24.1
psycopg2-binary==2.9.9
pyodbc==4.0.39
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
tenacity==9.0.0
typing_extensions==4.12.2
tzdata==2024.1
Unidecode==1.3.8
urllib3==2.2.3
